Begin decoupling stream scanning from Application scope
